Numerical Model Object Detection/requirements.txt
pillow
matplotlib
numpy
colorgram.py
